More modifications for GROMACS
1) Additional accessor functions to support modifying subtasks.
2) Prevent arithmetic bugs by using int64 and double in Ratio
   and Range classes.
3) Change task-yielding policy to run subtasks even if they are
   not yet ready (lambda not yet assigned). This prevents race
   conditions changing the order in which tasks are executed.

+++ sts/range.cpp
@@ -1,7 +1,7 @@
 #include "range.h"
 
-Ratio operator*(Ratio r, int n) { return Ratio(r.nom_*n,r.denom_); }
-Ratio operator*(int n, Ratio r) { return r*n; }
+Ratio operator*(Ratio r, int64_t n) { return Ratio(r.nom_*n,r.denom_); }
+Ratio operator*(int64_t n, Ratio r) { return r*n; }
 Ratio operator*(Ratio r1, Ratio r2) {
    return Ratio(r1.nom_ * r2.nom_ , r1.denom_ * r2.denom_).reduce();
 }
